+++ turbo/apps/platform/src/views/okou-page/sidebar-dialogs.tsx
@@ -138,6 +138,7 @@ export function AgentDialogSearch({
         />
         {query && (
           <Button
+            showTooltip
             type="button"
             onClick={() => {
               return setQuery("");
@@ -235,6 +236,7 @@ function AgentCommandSearch({
         />
         {query && (
           <Button
+            showTooltip
             type="button"
             onClick={() => {
               return setQuery("");