+++ turbo/packages/core/CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+## [8.595.2](https://github.com/vm0-ai/vm0/compare/core-v8.595.1...core-v8.595.2) (2026-08-26)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @okouai/api-contracts bumped to 1.372.0
+
+## [8.595.1](https://github.com/vm0-ai/vm0/compare/core-v8.595.0...core-v8.595.1) (2026-08-26)
+
+
+### Refactoring
+
+* remove home growth entry feature switch ([#29579](https://github.com/vm0-ai/vm0/issues/29579)) ([a6e7c72](https://github.com/vm0-ai/vm0/commit/a6e7c72715914cfe0bb4a4e780eac3c0c0285d47))
+
+## [8.595.0](https://github.com/vm0-ai/vm0/compare/core-v8.594.4...core-v8.595.0) (2026-08-26)
+
+
+### Features
+
+* add gradient color themes ([#29449](https://github.com/vm0-ai/vm0/issues/29449)) ([5ce77fb](https://github.com/vm0-ai/vm0/commit/5ce77fbbd64f2bf6ebf991587aab295f2a67155e))
+* **core:** enable connector accounts for staff ([#29611](https://github.com/vm0-ai/vm0/issues/29611)) ([bac7e55](https://github.com/vm0-ai/vm0/commit/bac7e5573d2b0ee4a43d00a0c43c8ce9c84baf44))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @okouai/api-contracts bumped to 1.371.0
+
 ## [8.594.4](https://github.com/vm0-ai/vm0/compare/core-v8.594.3...core-v8.594.4) (2026-08-26)